Enum skeletal_animation::blend_tree::BlendTreeNode

pub enum BlendTreeNode<T: Transform> {
    LerpNode(Box<BlendTreeNode<T>>, Box<BlendTreeNode<T>>, ParamId),
    AdditiveNode(Box<BlendTreeNode<T>>, Box<BlendTreeNode<T>>, ParamId),
    ClipNode(ClipInstance<T>),
}

Runtime representation of a blend tree.

Variants

LerpNode

Pose output is linear blend between the output of two child BlendTreeNodes, with blend factor according the paramater value for name ParamId

AdditiveNode

Pose output is additive blend between the output of two child BlendTreeNodes, with blend factor according the paramater value for name ParamId

ClipNode

Pose output is from an animation ClipInstance

Methods

impl<T: Transform> BlendTreeNode<T>

fn from_def(def: BlendTreeNodeDef, animations: &HashMap<ClipId, Rc<AnimationClip<T>>>) -> BlendTreeNode<T>

Initialize a new BlendTreeNode from a BlendTreeNodeDef and a mapping from animation names to AnimationClip

Arguments

def - The root BlendTreeNodeDef

animations - A mapping from ClipIds to shared AnimationClip instances

fn get_playback_length(&self, params: &HashMap<String, f32>) -> f32

Return the playback duration of the composite animation produced by this node and its subtree

fn set_playback_rate(&mut self, global_time: f32, rate: f32)

Set the playback rate of the composite animation produced by this node and its subtree

TODO - determine how / if we should handle setting the playback rate of a LerpNode / AdditiveNode

fn synchronize_subtree(&mut self, global_time: f32, params: &HashMap<String, f32>)

Adjust playback rates of subtree blend nodes so that playback lengths match. For blending looping animations with different durations.

fn get_output_pose(&mut self, time: f32, params: &HashMap<String, f32>, output_poses: &mut [T])

Get the output skeletal pose for this node and the given time and parameters

Arguments

time - The time to sample from any AnimationClips

params - A mapping from ParamIds to their current parameter values

output_poses - The output array slice of joint transforms that will be populated according to the defined output for this BlendTreeNode