Struct skeletal_animation::blend_tree::AdditiveAnimNode

pub struct AdditiveAnimNode {
    // some fields omitted
}

An AnimNode where pose output is additive blend with output of additive_input added to base_input, with blend factor according to value of blend_param