Module skeletal_animation::controller

Structs

AnimationController
A runtime representation of an Animation State Machine, consisting of one or more AnimationStates connected by AnimationTransitions, where the output animation pose depends on the current state or any active transitions between states.

AnimationControllerDef
Definition struct for an AnimationController, which can be deserialized from JSON and converted to an AnimationController instance at runtime

AnimationState
A state that an AnimationController can be in, consisting of a blend tree and a collection of transitions to other states

AnimationStateDef
Definition struct for an AnimationState, which can be deserialized from JSON and converted to an AnimationState instance at runtime

AnimationTransition
Representation of a state transition to a target state, with a condition and a duration

TransitionCondition
Representation of a condition to check for an AnimationTransition

Enums

Operator