Struct skeletal_animation::controller::AnimationController

pub struct AnimationController {
    // some fields omitted
}

A runtime representation of an Animation State Machine, consisting of one or more AnimationStates connected by AnimationTransitions, where the output animation pose depends on the current state or any active transitions between states.

Methods

impl AnimationController

fn new(controller_def: AnimationControllerDef, skeleton: Rc<Skeleton>, animations: &HashMap<ClipId, Rc<AnimationClip>>) -> AnimationController

Create an AnimationController instance from its definition, the desired skeleton, and a collection of currently loaded animation clips.

fn update(&mut self, delta_time: f64)

Update the controller's local clock with the given time delta

fn set_playback_speed(&mut self, speed: f64)

Set the playback speed for the controller

fn set_param_value(&mut self, name: &str, value: f32)

Set the value for the given controller parameter

fn get_param_value(&self, name: &str) -> f32

Return the value for the given controller parameter

fn get_parameters(&self) -> &HashMap<String, f32>

Return a read-only reference to the controller parameter map

fn get_output_pose(&mut self, ext_dt: f64, output_poses: &mut [Matrix4<f32>])

Calculate global skeletal joint poses for the given time since last update