Struct skeletal_animation::controller::AnimationControllerDef

pub struct AnimationControllerDef {
    pub parameters: Vec<String>,
    pub states: Vec<AnimationStateDef>,
    pub initial_state: String,
}

Definition struct for an AnimationController, which can be deserialized from JSON and converted to an AnimationController instance at runtime

Fields

parameters

Declaration list of all parameters that are used by the AnimationController, including state transition conditions and blend tree parameters

states

List of animation state definitions

initial_state

The name of the state that the AnimationController should start in