Struct skeletal_animation::controller::AnimationState

pub struct AnimationState<T: Transform> {
    pub blend_tree: AnimBlendTree<T>,
    pub transitions: Vec<AnimationTransition>,
}

A state that an AnimationController can be in, consisting of a blend tree and a collection of transitions to other states

Fields

blend_tree

The blend tree used to determine the final blended pose for this state

transitions

Transitions from this state to other AnimationStates