Enum skeletal_animation::blend_tree::BlendTreeNode

pub enum BlendTreeNode {
    LerpNode(Box<BlendTreeNode>, Box<BlendTreeNode>, ParamId),
    ClipNode(Rc<AnimationClip>),
}

Runtime representation of a blend tree.

Variants

LerpNode

Pose output is linearly blend between the output of two child BlendTreeNodes, with blend factor according the paramater value for name ParamId

ClipNode

Pose output is from an AnimationClip

Methods

impl BlendTreeNode

fn from_def(def: BlendTreeNodeDef, animations: &HashMap<ClipId, Rc<AnimationClip>>) -> BlendTreeNode

Initialize a new BlendTreeNode from a BlendTreeNodeDef and a mapping from animation names to AnimationClip

Arguments

def - The root BlendTreeNodeDef

animations - A mapping from ClipIds to shared AnimationClip instances

fn get_output_pose(&self, time: f32, params: &HashMap<String, f32>, output_poses: &mut [Transform])

Get the output skeletal pose for this node and the given time and parameters

Arguments

time - The time to sample from any AnimationClips

params - A mapping from ParamIds to their current parameter values

output_poses - The output array slice of joint transforms that will be populated according to the defined output for this BlendTreeNode